+++ module/web/__index__.ts
@@ -36,6 +36,7 @@ export * from './src/interceptor/get-cache.ts';
 export * from './src/interceptor/logging.ts';
 export * from './src/interceptor/respond.ts';
 
+export * from './src/util/body.ts';
 export * from './src/util/endpoint.ts';
 export * from './src/util/mime.ts';
 export * from './src/util/cookie.ts';
